backend/requirements.txt
setuptools>=70.0.0
wheel
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
google-generativeai==0.3.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
python-dotenv==1.0.0
redis==5.0.1
aiofiles==23.2.1
httpx==0.25.2
dateutils==0.6.12
email-validator==2.1.0
passlib[bcrypt]==1.7.4
cryptography>=3.4.0
PyJWT==2.8.0
nltk==3.8.1
numpy==1.26.4
pandas==2.1.4
